Make RequestId generic
Replace previous RequestId (alias to u64) with a trait with requirements
suited for use as HashMap key. Clone is preferred to Copy as there were
mentions of protocols that need arbitary strings as request identifiers.

`RequestId`s are generated using `RequestIdSource` trait. Values of
`RequestIdSource` are created using `NewRequestIdSource` which is also a
requirement of `RequestId`, avoiding the need for another RequestId
related bound in ClientProto or ServerProto definitions.